🎨 Move prepare path to dart format
